Move eyecandy to pkg and require to cmd/hypper/
