+++ pkg/eyecandy/colors.go
